wip: split campaign growth state contract
